# BoxedWrapper optimizations