```
           django_template3.py             2.8s (4)             2.8s (4)  -0.2%
                 pyxl_bench.py             2.2s (4)             2.2s (4)  -0.7%
     sqlalchemy_imperative2.py             2.8s (4)             2.7s (4)  -1.0%
                       geomean                 2.6s                 2.6s  -0.6%
```