@kmod 2015-08-31
nice :)  not a big deal either way, but is there a reason to have the 1-arg version take the extra always-NULL parameter?