@undingen 2015-09-01
I removed the unused always zero arg - should now be ready for merging