@undingen 2015-09-01
Oh I forgot that I now have commit access, I merged this one in because removing the NULL arg was pretty trivial change and you already reviewed the other changes before - hope that's fine.